added maps to navigation
