+++ client/js/screens/FindLocker/FindLocker.js
@@ -1,10 +1,28 @@
-import React from "react";
-import { Text, View } from "react-native";
+import React, { Component } from "react";
+import MapViewComponent from "../../components/MapView";
+import { View } from "react-native";
+import GoogleMapsSearch from "../../components/GoogleMapSearch";
+// import { SearchContext } from "../../components/GoogleMapSearch";
+import GooglePlaceContext from "../../context/GooglePlaceContext";
+import styles from "./styles";
 
-const FindLocker = ({ params }) => (
-  <View>
-    <Text>This is the Find Locker Screen</Text>
-  </View>
-);
+class FindLocker extends Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    return (
+      <GooglePlaceContext.Consumer>
+        {({ location, updateLocation }) => (
+          <View style={styles.map}>
+            <GoogleMapsSearch updateContextLocation={updateLocation} />
+            <MapViewComponent searchedLocation={location} />
+          </View>
+        )}
+      </GooglePlaceContext.Consumer>
+    );
+  }
+}
 
 export default FindLocker;
