+++ client/js/screens/FindLocker/FindLockerContainer.js
@@ -1,13 +1,15 @@
 import React, { Component } from "react";
 import FindLocker from "./FindLocker";
+import { View } from "react-native";
+import styles from "./styles";
 
-export default class FindLockerContainer extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
+class FindLockerContainer extends Component {
+  static navigationOptions = {
+    title: "Find Spot"
+  };
   render() {
     return <FindLocker />;
   }
 }
+
+export default FindLockerContainer;
