+++ client/js/screens/FindLocker/styles.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  text: {
+    position: "relative",
+    top: "50%",
+    textAlign: "center"
+  },
+  map: {
+    height: "100%"
+  }
+});
+
+export default styles;
