--- client/js/screens/FindSpot/FindSpotContainer.js
@@ -1,15 +1,0 @@
-import React, { Component } from "react";
-import FindSpot from "./FindSpot";
-import { View } from "react-native";
-import styles from "./styles";
-
-class FindSpotContainer extends Component {
-  static navigationOptions = {
-    title: "Find Spot"
-  };
-  render() {
-    return <FindSpot />;
-  }
-}
-
-export default FindSpotContainer;
