--- client/js/screens/FindSpot/index.js
@@ -1,3 +1,0 @@
-import FindSpotContainer from "./FindSpotContainer";
-
-export default FindSpotContainer;
