--- client/js/screens/FindSpot/styles.js
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  text: {
-    position: "relative",
-    top: "50%",
-    textAlign: "center"
-  },
-  map: {
-    height: "100%"
-  }
-});
-
-export default styles;
